+++ Xqueue/Maileon/Api/TestMarkAbandonedCartsInterface.php
@@ -1,0 +1,16 @@
+<?php
+ 
+namespace Xqueue\Maileon\Api;
+ 
+interface TestMarkAbandonedCartsInterface
+{
+    /**
+     * Test mark abandoned carts method
+     *
+     * @param string $token
+     *
+     * @return string
+     */
+ 
+    public function testMarkAbandonedCarts($token);
+}