+++ Xqueue/Maileon/Model/Order/Email/CreditmemoSender.php
@@ -179,13 +179,11 @@ class CreditmemoSender extends MaileonSender
         if ($contactCreated) {
             $transactionCreate = new TransactionCreate($this->pluginConfig['maileonApiKey'], 'no');
 
-            $transactionCreate->sendTransaction(
+            return $transactionCreate->sendTransaction(
                 $order->getCustomerEmail(),
                 'magento_order_creditmemo_v1',
                 $this->createTransactionContent($order, $creditmemo)
             );
-
-            return true;
         }
 
         return false;